add reset command

+++ src/Projectors/ProjectsEvents.php
@@ -5,6 +5,7 @@ namespace Spatie\EventProjector\Projectors;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Concerns\HasEvents;
 use Spatie\EventProjector\EventHandlers\HandlesEvents;
+use Spatie\EventProjector\Exceptions\CouldNotResetProjector;
 use Spatie\EventProjector\Models\StoredEvent;
 use Spatie\EventProjector\Models\ProjectorStatus;
 
@@ -46,8 +47,14 @@ trait ProjectsEvents
         return $this->getStatus()->updated_at;
     }
 
-    public function resetStatus()
+    public function reset()
     {
+        if (! method_exists($this, 'resetState')) {
+            throw CouldNotResetProjector::doesNotHaveResetStateMethod($this);
+        }
+
+        $this->resetState();
+
         $this->getStatus()->delete();
     }
